Fact_Check_Assistant/src/frontend/requirements.txt
# Core Agent Framework
strands-agents==1.0.1
strands-agents-tools==0.2.1

# MCP (Model Context Protocol) Dependencies
mcp>=1.0.0
duckduckgo-mcp-server==0.1.1

# Streamlit Web Framework
streamlit>=1.28.0

# Video/Caption Processing
webvtt-py>=0.4.6

# Environment Management
python-dotenv>=1.0.0

# Interactive Python (for debugging)
ipython>=8.0.0

# AWS SDK (for Bedrock models)
boto3>=1.34.0
botocore>=1.34.0

# HTTP Client (for health checks and API calls)
requests>=2.31.0

# Package manager for MCP servers
uv>=0.1.0

# OpenTelemetry for observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Additional dependencies that may be required by strands-agents
pydantic>=2.0.0
typing-extensions>=4.8.0
